+++ apigw/src/enduser/__tests__/csrf.ts
@@ -5,7 +5,7 @@
 import { GatewayTester } from '../../shared/test/gateway-tester'
 import app from '../app'
 import { csrfCookieName } from '../../shared/middleware/csrf'
-import { AuthenticatedUser } from '../../shared/service/pis'
+import { AuthenticatedUser } from '../../shared/service-client'
 
 const mockUser: AuthenticatedUser = {
   id: '4f73e4f8-8759-46c6-9b9d-4da860138ce2',
