+++ apigw/src/shared/service-client.ts
@@ -101,3 +101,94 @@ export function createServiceRequestHeaders(
   }
   return headers
 }
+
+const NIL_UUID = '00000000-0000-0000-0000-000000000000'
+
+export type UserRole =
+  | 'ENDUSER'
+  | 'ADMIN'
+  | 'DIRECTOR'
+  | 'FINANCE_ADMIN'
+  | 'SERVICE_WORKER'
+  | 'STAFF'
+  | 'UNIT_SUPERVISOR'
+
+export interface AuthenticatedUser {
+  id: UUID
+  roles: UserRole[]
+}
+
+export interface EmployeeIdentityRequest {
+  aad?: string
+  firstName: string
+  lastName: string
+  email: string
+}
+
+export interface EmployeeResponse {
+  id: string
+  aad: string | null
+  firstName: string
+  lastName: string
+  email: string
+  created: string
+  updated: string
+}
+
+export interface PersonIdentityRequest {
+  socialSecurityNumber: string
+  customerId?: number
+  firstName: string
+  lastName: string
+  email: string
+  language?: string
+}
+
+export async function getOrCreateEmployee(
+  employee: EmployeeIdentityRequest
+): Promise<AuthenticatedUser> {
+  return client
+    .post<AuthenticatedUser>(`/employee/identity`, employee, {
+      headers: createServiceRequestHeaders(undefined, {
+        id: NIL_UUID,
+        roles: []
+      })
+    })
+    .then((response) => response.data)
+}
+
+export async function getEmployeeDetails(
+  req: express.Request,
+  employeeId: string
+) {
+  return client
+    .get<EmployeeResponse>(`/employee/${employeeId}`, {
+      headers: createServiceRequestHeaders(req)
+    })
+    .then((res) => {
+      return res.data
+    })
+}
+
+export async function getOrCreatePerson(
+  person: PersonIdentityRequest
+): Promise<AuthenticatedUser> {
+  return client
+    .post<AuthenticatedUser>(`/person/identity`, person, {
+      headers: createServiceRequestHeaders(undefined, {
+        id: NIL_UUID,
+        roles: []
+      })
+    })
+    .then((response) => response.data)
+}
+
+export async function getUserDetails(req: express.Request) {
+  return client
+    .get(`/persondetails/uuid/${req.user?.id}`, {
+      headers: createServiceRequestHeaders(req)
+    })
+    .then((res) => {
+      return res.data
+    })
+}
